refactor: move phone validation inside form helpers
